feat: 각 fragment에서 화면 감지 GA 설정
