+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/write/OfferingWriteFragment.kt
@@ -275,6 +275,14 @@ class OfferingWriteFragment : Fragment(), OnOfferingWriteClickListener {
         toast?.show()
     }
 
+    override fun onResume() {
+        super.onResume()
+        firebaseAnalyticsManager.logScreenView(
+            screenName = "OfferingWriteFragment",
+            screenClass = this::class.java.simpleName,
+        )
+    }
+
     override fun onDestroy() {
         super.onDestroy()
         _fragmentBinding = null
